Update SDK with structured content snapshots
